requirements/production.txt
Django==2.0.7
mysqlclient
django-filter
gunicorn
Pillow
django-dynamic-formsets
boto3
django-storages
django-formset-js-improved
django-ipware
django-registration-redux
pinax-messages
psycopg2
python-dateutil
sorl-thumbnail
django-amazon-ses
requests
spotipy